moved trainconfig and evalconfig into utils
